requirements_production.txt
git+https://github.com/botswana-harvard/django-crypto-fields.git@0.1.14#egg=django_crypto_fields
git+https://github.com/botswana-harvard/django-revision.git@0.1.14#egg=django-revision
git+https://github.com/botswana-harvard/edc-action-item.git@0.1.5#egg=edc-action-item
git+https://github.com/botswana-harvard/edc-appointment.git@0.2.20#egg=edc-appointment
git+https://github.com/botswana-harvard/edc-base.git@0.2.28#egg=edc-base
git+https://github.com/botswana-harvard/edc-call-manager.git@0.2.3#egg=edc-call-manager
git+https://github.com/botswana-harvard/edc-calendar.git@0.1.4#egg=edc-calendar
git+https://github.com/botswana-harvard/edc-constants.git@0.2.9#egg=edc-constants
git+https://github.com/botswana-harvard/edc-consent.git@0.2.10#egg=edc-consent
git+https://github.com/botswana-harvard/edc-dashboard.git@0.3.3#egg=edc-dashboard
git+https://github.com/botswana-harvard/edc-data-manager.git@0.1.8#egg=edc-data-manager
git+https://github.com/botswana-harvard/edc-device.git@0.2.7#egg=edc-device
git+https://github.com/botswana-harvard/edc-facility.git@0.1.4#egg=edc-facility
git+https://github.com/botswana-harvard/edc-fieldsets.git@0.1.2#egg=edc-fieldsets
git+https://github.com/botswana-harvard/edc-form-validators.git@0.1.3#egg=edc-form-validators
git+https://github.com/botswana-harvard/edc-identifier.git@0.2.12#egg=edc-identifier
git+https://github.com/botswana-harvard/edc-lab.git@0.2.23#egg=edc-lab
git+https://github.com/botswana-harvard/edc-lab-dashboard.git@0.1.11#egg=edc-lab-dashboard
git+https://github.com/botswana-harvard/edc-label.git@0.1.9#egg=edc-label
git+https://github.com/botswana-harvard/edc-locator.git@0.2.7#egg=edc-locator
git+https://github.com/botswana-harvard/edc-list-data.git@0.1.1#egg=edc-list-data
git+https://github.com/botswana-harvard/edc-metadata.git@0.2.13#egg=edc-metadata
git+https://github.com/botswana-harvard/edc-metadata-rules.git@0.1.8#egg=edc-metadata-rules
git+https://github.com/botswana-harvard/edc-model-admin.git@0.1.2#egg=edc-model-admin
git+https://github.com/botswana-harvard/edc-model-wrapper.git@0.1.7#egg=edc-model-wrapper
git+https://github.com/botswana-harvard/edc-navbar.git@0.1.5#egg=edc-navbar
git+https://github.com/botswana-harvard/edc-prn.git@0.1.4#egg=edc-prn
git+https://github.com/botswana-harvard/edc-protocol.git@0.1.5#egg=edc-protocol
git+https://github.com/botswana-harvard/edc-reference.git@0.1.14#egg=edc-reference
git+https://github.com/botswana-harvard/edc-registration.git@0.2.11#egg=edc-registration
git+https://github.com/botswana-harvard/edc-reports.git@0.1.1#egg=edc-reports
git+https://github.com/botswana-harvard/edc-search.git@0.1.6#egg=edc-search
git+https://github.com/botswana-harvard/edc-senaite-interface.git@0.1.9#egg=edc-senaite-interface
git+https://github.com/botswana-harvard/edc-sms.git@0.1.1#egg=edc-sms
git+https://github.com/botswana-harvard/edc-subject-dashboard.git@0.2.6#egg=edc-subject-dashboard
git+https://github.com/botswana-harvard/edc-timepoint.git@0.1.3#egg=edc-timepoint
git+https://github.com/botswana-harvard/edc-visit-schedule.git@0.2.13#egg=edc-visit-schedule
git+https://github.com/botswana-harvard/edc-visit-tracking.git@0.2.13#egg=edc-visit-tracking
git+https://github.com/covid19-vaccine/esr21-subject.git@0.2.16#egg=esr21-subject
git+https://github.com/covid19-vaccine/esr21-prn.git@0.1.3#egg=esr21-prn
git+https://github.com/botswana-harvard/edc-rest.git@0.1.2#egg=edc-rest
git+https://github.com/botswana-harvard/edc-sync.git@0.2.29#egg=edc-sync
git+https://github.com/botswana-harvard/edc-sync-files.git@0.1.20#egg=edc-sync-files
git+https://github.com/covid19-vaccine/esr21-subject-validation.git@0.1.10#egg=esr21-subject-validation
git+https://github.com/covid19-vaccine/esr21-dashboard.git@0.1.4#egg=esr21-dashboard
git+https://github.com/covid19-vaccine/esr21-export.git@0.1.13#egg=esr21-export
git+https://github.com/covid19-vaccine/esr21-follow.git@0.1.6#egg=esr21-follow
git+https://github.com/covid19-vaccine/esr21-visit-schedule.git@0.1.10#egg=esr21-visit-schedule
git+https://github.com/covid19-vaccine/esr21-metadata-rules.git@0.1.8#egg=esr21-metadata-rules
git+https://github.com/covid19-vaccine/esr21-reference.git@0.1.5#egg=esr21-reference
git+https://github.com/covid19-vaccine/esr21-labs.git@0.1.14#egg=esr21-labs
git+https://github.com/covid19-vaccine/esr21-reports.git@0.1.5#egg=esr21-reports
mysqlclient
django-countries
django-multiselectfield
notebook
pandas
django-crispy-forms
xlwt
django-q
django-session-timeout
openpyxl
django-chartjs
django-pandas
flexitext
matplotlib